[feature]: WORKING (Session 04) Updates Mymap insert to account for tab in datasets.

+++ include/MyMap.hpp
@@ -25,6 +25,6 @@ class MyMap{
         void setLengths(int id, int length){lengths[id] = length;}
         int getLengths(int id){return lengths[id];}
         void print(int i){std::cout << " " << documents[i] << "\n";}
-        const int getsize(){return documents.size();}
+        const int getsize(){return (int)documents.size();}
 };
 #endif // MYMAP_H
